[Silver III] Title: 피보나치 함수, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1003. 피보나치 함수/피보나치 함수.cc
@@ -8,7 +8,7 @@ using namespace std;
 
 int t, n;
 
-pair<long long,long long> dp[41];
+pair<int,int> dp[41];
 int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
